requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tree-sitter>=0.22.0
tree-sitter-java>=0.21.0
rank-bm25>=0.2.2
pydantic>=2.0.0
numpy>=1.24.0
